Move classes to his own files
Some static variables and functions of become member of classes but
in overal there is no logical changes.
