+++ comm.h
@@ -22,86 +22,46 @@
 #include <errno.h>
 #include <inttypes.h>
 
-#include <mavlink.h>
+#include "macro.h"
 
 struct buffer {
     unsigned int len;
     uint8_t *data;
 };
 
-class Endpoint {
-public:
-    Endpoint(const char *name, bool crc_check_enabled);
-    virtual ~Endpoint();
-
-    int read_msg(struct buffer *pbuf, int *target_system, int *target_component);
-    void print_statistics();
-    virtual int write_msg(const struct buffer *pbuf) = 0;
-    virtual int flush_pending_msgs() = 0;
-
-    uint8_t get_system_id() { return system_id; }
-    uint8_t get_component_id() { return component_id; }
-
-    struct buffer rx_buf;
-    struct buffer tx_buf;
-    int fd = -1;
-
-protected:
-    virtual ssize_t _read_msg(uint8_t *buf, size_t len) = 0;
-    bool _check_crc(const mavlink_msg_entry_t *msg_entry);
-
-    const char *_name;
-    size_t _last_packet_len = 0;
-
-    uint32_t _read_crc_errors = 0;
-    uint32_t _read_total = 0;
-    uint32_t _write_total = 0;
-    const bool _crc_check_enabled;
-
-    uint8_t system_id = 0;
-    uint8_t component_id = 0;
+/*
+ * mavlink 2.0 packet in its wire format
+ *
+ * Packet size:
+ *      sizeof(mavlink_router_mavlink2_header)
+ *      + payload length
+ *      + 2 (checksum)
+ *      + signature (0 if not signed)
+ */
+struct _packed_ mavlink_router_mavlink2_header {
+    uint8_t magic;
+    uint8_t payload_len;
+    uint8_t incompat_flags;
+    uint8_t compat_flags;
+    uint8_t seq;
+    uint8_t sysid;
+    uint8_t compid;
+    uint32_t msgid : 24;
 };
 
-class UartEndpoint : public Endpoint {
-public:
-    UartEndpoint() : Endpoint{"UART", true} { }
-    virtual ~UartEndpoint() { }
-    int write_msg(const struct buffer *pbuf) override;
-    int flush_pending_msgs() override { return -ENOSYS; }
-
-    int open(const char *path, speed_t baudrate);
-protected:
-    ssize_t _read_msg(uint8_t *buf, size_t len) override;
-};
-
-class UdpEndpoint : public Endpoint {
-public:
-    UdpEndpoint();
-    virtual ~UdpEndpoint() { }
-
-    int write_msg(const struct buffer *pbuf) override;
-    int flush_pending_msgs() override { return -ENOSYS; }
-
-    int open(const char *ip, unsigned long port, bool bind = false);
-
-    struct sockaddr_in sockaddr;
-
-protected:
-    ssize_t _read_msg(uint8_t *buf, size_t len) override;
-};
-
-class TcpEndpoint : public Endpoint {
-public:
-    TcpEndpoint();
-    virtual ~TcpEndpoint();
-
-    int accept(int listener_fd);
-
-    int write_msg(const struct buffer *pbuf) override;
-    int flush_pending_msgs() override { return -ENOSYS; }
-
-    struct sockaddr_in sockaddr;
-
-protected:
-    ssize_t _read_msg(uint8_t *buf, size_t len) override;
+/*
+ * mavlink 1.0 packet in its wire format
+ *
+ * Packet size:
+ *      sizeof(mavlink_router_mavlink1_header)
+ *      + payload length
+ *      + 2 (checksum)
+ */
+struct _packed_ mavlink_router_mavlink1_header {
+    uint8_t magic;
+    uint8_t payload_len;
+    uint8_t seq;
+    uint8_t sysid;
+    uint8_t compid;
+    uint8_t msgid;
 };
